Update TestStoreFactory to return MutablePrunableStore
