+++ ethereum/datastructures/src/testFixtures/java/tech/pegasys/teku/datastructures/forkchoice/TestStoreFactory.java
@@ -13,21 +13,18 @@
 
 package tech.pegasys.teku.datastructures.forkchoice;
 
+import static com.google.common.base.Preconditions.checkArgument;
 import static tech.pegasys.teku.util.config.Constants.SECONDS_PER_SLOT;
 
 import com.google.common.primitives.UnsignedLong;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.Optional;
-import java.util.Set;
 import org.apache.tuweni.bytes.Bytes32;
 import tech.pegasys.teku.bls.BLSSignature;
 import tech.pegasys.teku.datastructures.blocks.BeaconBlock;
 import tech.pegasys.teku.datastructures.blocks.SignedBeaconBlock;
-import tech.pegasys.teku.datastructures.blocks.SignedBlockAndState;
 import tech.pegasys.teku.datastructures.state.BeaconState;
 import tech.pegasys.teku.datastructures.state.Checkpoint;
-import tech.pegasys.teku.datastructures.state.CheckpointAndBlock;
 import tech.pegasys.teku.datastructures.util.BeaconStateUtil;
 import tech.pegasys.teku.datastructures.util.DataStructureUtil;
 import tech.pegasys.teku.util.config.Constants;
@@ -35,22 +32,20 @@ import tech.pegasys.teku.util.config.Constants;
 public class TestStoreFactory {
   private final DataStructureUtil dataStructureUtil = new DataStructureUtil();
 
-  public ReadOnlyStore createGenesisStore() {
-    final BeaconState genesisStore = createRandomGenesisState();
-    return getForkChoiceStore(StoreImpl::new, genesisStore);
+  public MutablePrunableStore createGenesisStore() {
+    final BeaconState genesisState = createRandomGenesisState();
+    return getForkChoiceStore(genesisState);
   }
 
-  public ReadOnlyStore createGenesisStore(final BeaconState genesisState) {
-    return getForkChoiceStore(StoreImpl::new, genesisState);
+  public MutablePrunableStore createGenesisStore(final BeaconState genesisState) {
+    checkArgument(
+        genesisState.getSlot().equals(UnsignedLong.valueOf(Constants.GENESIS_SLOT)),
+        "Genesis state has invalid slot.");
+    return getForkChoiceStore(genesisState);
   }
 
-  public MutableStore createMutableGenesisStore() {
-    final BeaconState genesisStore = createRandomGenesisState();
-    return getForkChoiceStore(MutableStoreImpl::new, genesisStore);
-  }
-
-  public MutableStore createEmptyStore() {
-    return new MutableStoreImpl(
+  public MutablePrunableStore createEmptyStore() {
+    return new TestStoreImpl(
         UnsignedLong.ZERO,
         UnsignedLong.ZERO,
         null,
@@ -62,8 +57,7 @@ public class TestStoreFactory {
         new HashMap<>());
   }
 
-  private <T extends ReadOnlyStore> T getForkChoiceStore(
-      StoreConstructor<T> storeConstructor, final BeaconState anchorState) {
+  private MutablePrunableStore getForkChoiceStore(final BeaconState anchorState) {
     final BeaconBlock anchorBlock = new BeaconBlock(anchorState.hash_tree_root());
     final SignedBeaconBlock signedAnchorBlock =
         new SignedBeaconBlock(anchorBlock, BLSSignature.empty());
@@ -80,7 +74,7 @@ public class TestStoreFactory {
     block_states.put(anchorRoot, anchorState);
     checkpoint_states.put(anchorCheckpoint, anchorState);
 
-    return storeConstructor.create(
+    return new TestStoreImpl(
         anchorState
             .getGenesis_time()
             .plus(UnsignedLong.valueOf(SECONDS_PER_SLOT).times(anchorState.getSlot())),
@@ -96,212 +90,5 @@ public class TestStoreFactory {
 
   private BeaconState createRandomGenesisState() {
     return dataStructureUtil.randomBeaconState(UnsignedLong.valueOf(Constants.GENESIS_SLOT));
-  }
-
-  private static class StoreImpl implements ReadOnlyStore {
-    protected UnsignedLong time;
-    protected UnsignedLong genesis_time;
-    protected Checkpoint justified_checkpoint;
-    protected Checkpoint finalized_checkpoint;
-    protected Checkpoint best_justified_checkpoint;
-    protected Map<Bytes32, SignedBeaconBlock> blocks;
-    protected Map<Bytes32, BeaconState> block_states;
-    protected Map<Checkpoint, BeaconState> checkpoint_states;
-    protected Map<UnsignedLong, VoteTracker> votes;
-
-    StoreImpl(
-        final UnsignedLong time,
-        final UnsignedLong genesis_time,
-        final Checkpoint justified_checkpoint,
-        final Checkpoint finalized_checkpoint,
-        final Checkpoint best_justified_checkpoint,
-        final Map<Bytes32, SignedBeaconBlock> blocks,
-        final Map<Bytes32, BeaconState> block_states,
-        final Map<Checkpoint, BeaconState> checkpoint_states,
-        final Map<UnsignedLong, VoteTracker> votes) {
-      this.time = time;
-      this.genesis_time = genesis_time;
-      this.justified_checkpoint = justified_checkpoint;
-      this.finalized_checkpoint = finalized_checkpoint;
-      this.best_justified_checkpoint = best_justified_checkpoint;
-      this.blocks = blocks;
-      this.block_states = block_states;
-      this.checkpoint_states = checkpoint_states;
-      this.votes = votes;
-    }
-
-    @Override
-    public UnsignedLong getTime() {
-      return time;
-    }
-
-    @Override
-    public UnsignedLong getGenesisTime() {
-      return genesis_time;
-    }
-
-    @Override
-    public Checkpoint getJustifiedCheckpoint() {
-      return justified_checkpoint;
-    }
-
-    @Override
-    public Checkpoint getFinalizedCheckpoint() {
-      return finalized_checkpoint;
-    }
-
-    @Override
-    public CheckpointAndBlock getFinalizedCheckpointAndBlock() {
-      final SignedBeaconBlock block = getSignedBlock(finalized_checkpoint.getRoot());
-      return new CheckpointAndBlock(finalized_checkpoint, block);
-    }
-
-    @Override
-    public UnsignedLong getLatestFinalizedBlockSlot() {
-      return blocks.get(finalized_checkpoint.getRoot()).getSlot();
-    }
-
-    @Override
-    public SignedBlockAndState getLatestFinalizedBlockAndState() {
-      final SignedBeaconBlock block = getSignedBlock(finalized_checkpoint.getRoot());
-      final BeaconState state = getBlockState(finalized_checkpoint.getRoot());
-      return new SignedBlockAndState(block, state);
-    }
-
-    @Override
-    public Checkpoint getBestJustifiedCheckpoint() {
-      return best_justified_checkpoint;
-    }
-
-    @Override
-    public BeaconBlock getBlock(final Bytes32 blockRoot) {
-      return blocks.get(blockRoot).getMessage();
-    }
-
-    @Override
-    public SignedBeaconBlock getSignedBlock(final Bytes32 blockRoot) {
-      return blocks.get(blockRoot);
-    }
-
-    @Override
-    public Optional<SignedBlockAndState> getBlockAndState(final Bytes32 blockRoot) {
-      final SignedBeaconBlock block = getSignedBlock(blockRoot);
-      final BeaconState state = getBlockState(blockRoot);
-      if (block == null || state == null) {
-        return Optional.empty();
-      }
-      return Optional.of(new SignedBlockAndState(block, state));
-    }
-
-    @Override
-    public boolean containsBlock(final Bytes32 blockRoot) {
-      return blocks.containsKey(blockRoot);
-    }
-
-    @Override
-    public Set<Bytes32> getBlockRoots() {
-      return blocks.keySet();
-    }
-
-    @Override
-    public BeaconState getBlockState(final Bytes32 blockRoot) {
-      return block_states.get(blockRoot);
-    }
-
-    @Override
-    public Optional<BeaconState> getCheckpointState(final Checkpoint checkpoint) {
-      return Optional.ofNullable(checkpoint_states.get(checkpoint));
-    }
-
-    @Override
-    public Set<UnsignedLong> getVotedValidatorIndices() {
-      return votes.keySet();
-    }
-  }
-
-  private static class MutableStoreImpl extends StoreImpl implements MutableStore {
-
-    MutableStoreImpl(
-        final UnsignedLong time,
-        final UnsignedLong genesis_time,
-        final Checkpoint justified_checkpoint,
-        final Checkpoint finalized_checkpoint,
-        final Checkpoint best_justified_checkpoint,
-        final Map<Bytes32, SignedBeaconBlock> blocks,
-        final Map<Bytes32, BeaconState> block_states,
-        final Map<Checkpoint, BeaconState> checkpoint_states,
-        final Map<UnsignedLong, VoteTracker> votes) {
-      super(
-          time,
-          genesis_time,
-          justified_checkpoint,
-          finalized_checkpoint,
-          best_justified_checkpoint,
-          blocks,
-          block_states,
-          checkpoint_states,
-          votes);
-    }
-
-    @Override
-    public void putBlockAndState(final SignedBeaconBlock block, final BeaconState state) {
-      blocks.put(block.getRoot(), block);
-      block_states.put(block.getRoot(), state);
-    }
-
-    @Override
-    public void putBlockAndState(final SignedBlockAndState blockAndState) {
-      blocks.put(blockAndState.getRoot(), blockAndState.getBlock());
-      block_states.put(blockAndState.getRoot(), blockAndState.getState());
-    }
-
-    @Override
-    public void setTime(final UnsignedLong time) {
-      this.time = time;
-    }
-
-    @Override
-    public void setGenesis_time(final UnsignedLong genesis_time) {
-      this.genesis_time = genesis_time;
-    }
-
-    @Override
-    public void setJustifiedCheckpoint(final Checkpoint justified_checkpoint) {
-      this.justified_checkpoint = justified_checkpoint;
-    }
-
-    @Override
-    public void setFinalizedCheckpoint(final Checkpoint finalized_checkpoint) {
-      this.finalized_checkpoint = finalized_checkpoint;
-    }
-
-    @Override
-    public void setBestJustifiedCheckpoint(final Checkpoint best_justified_checkpoint) {
-      this.best_justified_checkpoint = best_justified_checkpoint;
-    }
-
-    @Override
-    public VoteTracker getVote(final UnsignedLong validatorIndex) {
-      VoteTracker vote = votes.get(validatorIndex);
-      if (vote == null) {
-        vote = VoteTracker.Default();
-      }
-      this.votes.put(validatorIndex, vote);
-      return vote;
-    }
-  }
-
-  @FunctionalInterface
-  private interface StoreConstructor<T extends ReadOnlyStore> {
-    T create(
-        final UnsignedLong time,
-        final UnsignedLong genesis_time,
-        final Checkpoint justified_checkpoint,
-        final Checkpoint finalized_checkpoint,
-        final Checkpoint best_justified_checkpoint,
-        final Map<Bytes32, SignedBeaconBlock> blocks,
-        final Map<Bytes32, BeaconState> block_states,
-        final Map<Checkpoint, BeaconState> checkpoint_states,
-        final Map<UnsignedLong, VoteTracker> votes);
   }
 }
